todo details working

+++ src/Screens/AllProperties.tsx
@@ -25,9 +25,12 @@ function AllProperties() {
 			<SwipeListView
 				data={appContext.Properties}
 				rightOpenValue={-50}
-				renderHiddenItem={({ item }) => (
-					<TrashButton onPress={()=>handleDeleteProperty(item.PropertyId)}/>
-				)}
+				renderHiddenItem={({ item, index }) => {
+					if (appContext.Properties[index].isTenant) return null;
+					return (
+						<TrashButton onPress={() => handleDeleteProperty(item.PropertyId)} />
+					);
+				}}
 				renderItem={({ item }) => (
 					<PropertyCard property={item} />
 				)}
